refactor: move ssh/gateway models
