+++ internal/provider/model/connection/gateway.go
@@ -1,4 +1,4 @@
-package model
+package connection
 
 import "github.com/hashicorp/terraform-plugin-framework/types"
