+++ internal/provider/model/connection/ssl.go
@@ -1,4 +1,4 @@
-package model
+package connection
 
 import "github.com/hashicorp/terraform-plugin-framework/types"
